Install packages only when required

+++ common/devscripts/provisioning.sh
@@ -17,15 +17,17 @@ fi
 ############
 INSTALL_PACKAGES=${INSTALL_PACKAGES:-"git less zsh"}
 
-if command -v dnf; then
-    dnf install -y ${INSTALL_PACKAGES};
-    dnf clean all;
-elif command -v apt; then
-    apt update;
-    apt install -y ${INSTALL_PACKAGES};
-    rm -rf /var/lib/apt/lists/*;
-elif command -v apk; then
-    apk add --no-cache ${INSTALL_PACKAGES};
+if [ -n "${INSTALL_PACKAGES}" ]; then
+    if command -v dnf; then
+        dnf install -y ${INSTALL_PACKAGES};
+        dnf clean all;
+    elif command -v apt; then
+        apt update;
+        apt install -y ${INSTALL_PACKAGES};
+        rm -rf /var/lib/apt/lists/*;
+    elif command -v apk; then
+        apk add --no-cache ${INSTALL_PACKAGES};
+    fi
 fi
 
 ########
